requirements.txt
apache-airflow
pandas
numpy
pendulum
plotly
tabulate
schedule
setuptools